+++ packages/nuxt/src/module.ts
@@ -11,11 +11,11 @@ import YAML from 'yaml'
 import { version } from '../package.json'
 import { setupDevToolsUI } from './devtools'
 import { DEVTOOLS_MODULE_KEY, DEVTOOLS_MODULE_NAME } from './constants'
-import type { PergelOptions, ResolvedPergelOptions } from './runtime/core/types/module'
 import { useNitroImports, useNuxtImports } from './runtime/core/utils/useImports'
 import { setupPergel } from './runtime/core/setupPergel'
 import { generateReadmeYaml } from './runtime/core/utils/generateYaml'
 import { setupModules } from './runtime/core/setupModules'
+import type { PergelOptions, ResolvedPergelOptions } from './runtime/core/types/nuxtModule'
 
 export default defineNuxtModule<PergelOptions>({
   meta: {
