--- packages/nuxt/src/runtime/core/types/index.ts
@@ -1,2 +1,0 @@
-export type * from './module'
-export type * from './devtools'
