+++ packages/nuxt/src/runtime/core/utils/index.ts
@@ -1,5 +1,5 @@
 import { snakeCase } from 'scule'
-import type { EnvList } from '../types'
+import type { EnvList } from '../types/module'
 export { useModuleOptions } from './useModuleOptions'
 /**
  * Reformat source code for improved readability by adding appropriate newlines and indentation.
