+++ packages/nuxt/src/runtime/modules/S3/types.ts
@@ -1,8 +1,10 @@
-export interface S3ModuleOptions {
+import type { PergelModuleOptions, ResolvedPergelModuleOptions } from '../../core/types/module'
+
+export interface S3ModuleOptions extends PergelModuleOptions {
 
 }
 
-export interface S3ModuleRuntimeConfig {
+export interface S3ModuleRuntimeConfig extends ResolvedPergelModuleOptions {
   region: string
   endpoint: string
   accessKeyId: string
